#376 Add wrapper to ArchivedBoardsList
